requirements.txt
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
pyyaml==6.0.1
python-dotenv==1.0.0
anthropic==0.18.1
faker==22.6.0
tqdm==4.66.1
